+++ tests/test_app/urls.py
@@ -31,6 +31,7 @@ admin.site.has_permission = lambda r: setattr(r, "user", AccessUser()) or True  
 
 urlpatterns = [
     path("", base_template),
+    path("", include("test_app.performance.urls")),
     path("reactpy/", include("reactpy_django.http.urls")),
     path("admin/", admin.site.urls),
 ]